Add search history page and routing
