+++ README.md
@@ -2,6 +2,13 @@
 
 한국인을 위한 간단한 가격 비교 사이트입니다. Vite와 React로 제작되었습니다.
 
+## 엔드포인트
+
+| 경로 | 설명 |
+|------|------|
+| `/` | 상품 검색 및 가격 비교 페이지 |
+| `/history` | 검색한 내역을 볼 수 있는 히스토리 페이지 |
+
 ## Development
 
 Install dependencies and start the dev server:
